requirements.txt
annotated-types==0.6.0
anyio==4.2.0
azure-cognitiveservices-speech==1.34.0
BTrees==5.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
distro==1.9.0
fastapi==0.108.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
openai==1.6.1
packaging==23.2
persistent==5.1
pluggy==1.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
python-multipart==0.0.6
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tqdm==4.66.1
transaction==4.0
typing_extensions==4.9.0
uvicorn==0.25.0
zc.lockfile==3.0.post1
ZConfig==4.0
ZODB==5.8.1
zodbpickle==3.1
zope.interface==6.1
